requirements.txt
crewai>=0.28.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
langchain>=0.0.335
langchain-openai>=0.0.5
langchain-community>=0.0.10
pydantic>=2.5.0
ollama>=0.1.6
matplotlib>=3.7.0
plotly>=5.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
jinja2>=3.1.2
python-unsplash>=1.0.0
pillow>=10.0.0 